+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+* 1.0.0-next (@release_date@)
+
+ * optimized `Valid::andX()` and `Valid::orX()`
+ * optimized `Invalid::andX()` and `Invalid::orX()`
+
 * 1.0.0-beta2 (2015-04-13)
 
  * added `Selector`